+++ core/src/main/kotlin/net/corda/core/internal/TransactionUtils.kt
@@ -7,6 +7,10 @@ import net.corda.core.crypto.componentHash
 import net.corda.core.crypto.sha256
 import net.corda.core.flows.FlowLogic
 import net.corda.core.identity.Party
+import net.corda.core.node.services.AttachmentId
+import net.corda.core.node.services.AttachmentStorage
+import net.corda.core.node.services.vault.AttachmentQueryCriteria
+import net.corda.core.node.services.vault.Builder
 import net.corda.core.serialization.*
 import net.corda.core.transactions.*
 import net.corda.core.utilities.OpaqueBytes
@@ -179,4 +183,47 @@ fun FlowLogic<*>.checkParameterHash(networkParametersHash: SecureHash?) {
     //       We will never end up in perfect synchronization with all the nodes. However, network parameters update process
     //       lets us predict what is the reasonable time window for changing parameters on most of the nodes.
     //       For now we don't check whether the attached network parameters match the current ones.
+}
+
+private object AttachmentTrustedCache {
+    val cache: MutableMap<AttachmentId, Boolean> = createSimpleCache(100)
+}
+
+/**
+ * Establishes whether an attachment should be trusted. This logic is required in order to verify transactions, as transaction
+ * verification should only be carried out using trusted attachments.
+ *
+ * Attachments are trusted if one of the following is true:
+ *  - They are uploaded by a trusted uploader
+ *  - There is another attachment in the attachment store with the same contract classes and signed by the same keys that is trusted
+ */
+fun isAttachmentTrusted(attachment: Attachment, service: AttachmentStorage?): Boolean {
+    val trustedByUploader = when (attachment) {
+        is ContractAttachment -> isUploaderTrusted(attachment.uploader)
+        is AbstractAttachment -> isUploaderTrusted(attachment.uploader)
+        else -> false
+    }
+
+    if (trustedByUploader) return true
+
+    return if (service != null && (attachment is AbstractAttachment || attachment is ContractAttachment)) {
+        AttachmentTrustedCache.cache.computeIfAbsent(attachment.id) {
+            val signers = attachment.signerKeys
+            val contractClassCondition = if (attachment is ContractAttachment) {
+                val contractClasses = attachment.allContracts.toList()
+                Builder.equal(contractClasses)
+            } else {
+                null
+            }
+
+            val queryCriteria = AttachmentQueryCriteria.AttachmentsQueryCriteria(
+                    contractClassNamesCondition = contractClassCondition,
+                    signersCondition = Builder.equal(signers),
+                    uploaderCondition = Builder.`in`(TRUSTED_UPLOADERS)
+            )
+            service.queryAttachments(queryCriteria).isNotEmpty()
+        }
+    } else {
+        false
+    }
 }