+++ core/src/main/kotlin/net/corda/core/transactions/LedgerTransaction.kt
@@ -74,7 +74,7 @@ private constructor(
     private var componentGroups: List<ComponentGroup>? = null
     private var serializedInputs: List<SerializedStateAndRef>? = null
     private var serializedReferences: List<SerializedStateAndRef>? = null
-    private var attachmentTrustedFun: (Attachment) -> Boolean = { WireTransaction.isAttachmentTrusted(it, null) }
+    private var attachmentTrustedFun: (Attachment) -> Boolean = { isAttachmentTrusted(it, null) }
 
     init {
         if (timeWindow != null) check(notary != null) { "Transactions with time-windows must be notarised" }
@@ -143,7 +143,11 @@ private constructor(
     internal fun internalPrepareVerify(extraAttachments: List<Attachment>): Verifier {
         // Switch thread local deserialization context to using a cached attachments classloader. This classloader enforces various rules
         // like no-overlap, package namespace ownership and (in future) deterministic Java.
-        return AttachmentsClassLoaderBuilder.withAttachmentsClassloaderContext(this.attachments + extraAttachments, getParamsWithGoo(), id, isAttachmentTrusted = attachmentTrustedFun) { transactionClassLoader ->
+        return AttachmentsClassLoaderBuilder.withAttachmentsClassloaderContext(
+                this.attachments + extraAttachments,
+                getParamsWithGoo(),
+                id,
+                isAttachmentTrusted = attachmentTrustedFun) { transactionClassLoader ->
             // Create a copy of the outer LedgerTransaction which deserializes all fields inside the [transactionClassLoader].
             // Only the copy will be used for verification, and the outer shell will be discarded.
             // This artifice is required to preserve backwards compatibility.