+++ core/src/test/kotlin/net/corda/core/transactions/AttachmentsClassLoaderTests.kt
@@ -3,10 +3,10 @@ package net.corda.core.transactions
 import net.corda.core.contracts.Attachment
 import net.corda.core.contracts.Contract
 import net.corda.core.contracts.TransactionVerificationException
-import net.corda.core.crypto.Crypto
 import net.corda.core.crypto.SecureHash
 import net.corda.core.internal.declaredField
 import net.corda.core.internal.inputStream
+import net.corda.core.internal.isAttachmentTrusted
 import net.corda.core.node.NetworkParameters
 import net.corda.core.node.services.AttachmentId
 import net.corda.core.serialization.internal.AttachmentsClassLoader
@@ -22,7 +22,6 @@ import org.junit.Test
 import java.io.ByteArrayOutputStream
 import java.io.InputStream
 import java.net.URL
-import java.security.PublicKey
 import kotlin.test.assertFailsWith
 
 class AttachmentsClassLoaderTests {
@@ -44,7 +43,7 @@ class AttachmentsClassLoaderTests {
     private val networkParameters = testNetworkParameters()
     private fun make(attachments: List<Attachment>,
                      params: NetworkParameters = networkParameters): AttachmentsClassLoader {
-        return AttachmentsClassLoader(attachments, params, SecureHash.zeroHash, { WireTransaction.isAttachmentTrusted(it, storage) })
+        return AttachmentsClassLoader(attachments, params, SecureHash.zeroHash, { isAttachmentTrusted(it, storage) })
     }
 
     @Test