+++ node/src/main/kotlin/net/corda/node/migration/MigrationServicesForResolution.kt
@@ -6,6 +6,7 @@ import net.corda.core.cordapp.CordappProvider
 import net.corda.core.crypto.SecureHash
 import net.corda.core.internal.deserialiseComponentGroup
 import net.corda.core.internal.div
+import net.corda.core.internal.isAttachmentTrusted
 import net.corda.core.internal.readObject
 import net.corda.core.node.NetworkParameters
 import net.corda.core.node.ServicesForResolution
@@ -111,7 +112,7 @@ class MigrationServicesForResolution(
                     txAttachments,
                     networkParameters,
                     tx.id,
-                    { WireTransaction.isAttachmentTrusted(it, attachments) },
+                    { isAttachmentTrusted(it, attachments) },
                     cordappLoader.appClassLoader) {
                 deserialiseComponentGroup(tx.componentGroups, TransactionState::class, ComponentGroupEnum.OUTPUTS_GROUP, forceDeserialize = true)
             }