+++ docs/source/changelog.rst
@@ -4,6 +4,16 @@ Changelog
 Here's a summary of what's changed in each Corda release. For guidance on how to upgrade code from the previous
 release, see :doc:`app-upgrade-notes`.
 
+.. _changelog_v4.2:
+
+Version 4.2
+-----------
+
+* Contract attachments are now automatically whitelisted by the node if another contract attachment is present with the same contract classes,
+  signed by the same public keys, and uploaded by a trusted uploader. This allows the node to resolve transactions that use earlier versions
+  of a contract without having to manually install that version, provided a newer version is installed. Similarly, non-contract attachments
+  are whitelisted if another attachment is present on the node that is signed by the same public key.
+
 .. _changelog_v4.0:
 
 Version 4.0